+++ tests/ember_lts_2_18.js
@@ -56,7 +56,7 @@ test('2.18 LTS - map components', t => {
   );
 });
 
-test('2.18 LTS - look for unused components', t => {
+test('2.18 LTS - look for unused components and calculate stats', t => {
   let config = {
     appPath: '/test-apps/ember_lts_2_18/app/',
     ignore: ['app/templates/freestyle.hbs'],
@@ -77,6 +77,16 @@ test('2.18 LTS - look for unused components', t => {
 
   let expectedUnusedComponents = ['user/user-signature'];
 
+  let expectedStats = {
+    'user/user-avatar': { name: 'user/user-avatar', count: 1, curly: 1, angle: 0 },
+    'user/user-card': { name: 'user/user-card', count: 1, curly: 1, angle: 0 },
+    'user/user-info': { name: 'user/user-info', count: 1, curly: 1, angle: 0 },
+    'user/user-signature': { name: 'user/user-signature', count: 0, curly: 0, angle: 0 },
+    'x-button': { name: 'x-button', count: 1, curly: 1, angle: 0 },
+    'y-button': { name: 'y-button', count: 1, curly: 1, angle: 0 },
+    'z-button': { name: 'z-button', count: 0, curly: 0, angle: 0 },
+  };
+
   analyser.scanProject(config);
   analyser.respectWhitelist(config.whitelist);
 
@@ -86,4 +96,5 @@ test('2.18 LTS - look for unused components', t => {
     expectedUnusedComponents,
     'has proper list of unused components'
   );
+  t.deepEqual(analyser.stats, expectedStats, 'has properly calculated stats');
 });